refactor(api): move validate to src
